refactor: change directories
