+++ pkg/grpc/verb.go
@@ -1,16 +1,28 @@
 // common/fetchVerb.go
 
-package common
+package grpc
 
 import (
+	"context"
+	"crypto/tls"
 	"fmt"
 	"os"
 	"os/signal"
+	"path/filepath"
 	"sort"
 	"strings"
 	"time"
 
+	"github.com/cloudforet-io/cfctl/cmd/other"
+	"github.com/cloudforet-io/cfctl/pkg/format"
+	"github.com/cloudforet-io/cfctl/pkg/settings"
+	"github.com/jhump/protoreflect/grpcreflect"
 	"github.com/pterm/pterm"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials"
+	"google.golang.org/grpc/metadata"
+	"google.golang.org/grpc/reflection/grpc_reflection_v1alpha"
+	"gopkg.in/yaml.v3"
 
 	"github.com/spf13/cobra"
 )
@@ -34,7 +46,7 @@ type FetchOptions struct {
 // AddVerbCommands adds subcommands for each verb to the parent command
 func AddVerbCommands(parentCmd *cobra.Command, serviceName string, groupID string) error {
 	// Build the verb-resource map
-	verbResourceMap, err := BuildVerbResourceMap(serviceName)
+	verbResourceMap, err := buildVerbResourceMap(serviceName)
 	if err != nil {
 		return nil // Return nil to prevent Cobra from showing additional error messages
 	}
@@ -153,7 +165,7 @@ func AddVerbCommands(parentCmd *cobra.Command, serviceName string, groupID strin
 		verbCmd.Flags().BoolP("copy", "y", false, "Copy the output to the clipboard (copies any output format)")
 
 		// Set custom help function
-		verbCmd.SetHelpFunc(CustomVerbHelpFunc)
+		verbCmd.SetHelpFunc(format.CustomVerbHelpFunc)
 
 		// Update example for list command
 		if currentVerb == "list" {
@@ -164,6 +176,235 @@ func AddVerbCommands(parentCmd *cobra.Command, serviceName string, groupID strin
 	}
 
 	return nil
+}
+
+func buildVerbResourceMap(serviceName string) (map[string][]string, error) {
+	home, err := os.UserHomeDir()
+	if err != nil {
+		return nil, fmt.Errorf("failed to get home directory: %v", err)
+	}
+
+	config, err := settings.LoadSetting()
+	if err != nil {
+		return nil, fmt.Errorf("failed to load config: %v", err)
+	}
+
+	if config.Environment == "local" {
+		return handleLocalEnvironment(serviceName)
+	}
+
+	cacheDir := filepath.Join(home, ".cfctl", "cache", config.Environment)
+	cacheFile := filepath.Join(cacheDir, "verb_resources.yaml")
+
+	if info, err := os.Stat(cacheFile); err == nil {
+		if time.Since(info.ModTime()) < time.Hour {
+			data, err := os.ReadFile(cacheFile)
+			if err == nil {
+				var allServices map[string]map[string][]string
+				if err := yaml.Unmarshal(data, &allServices); err == nil {
+					if verbMap, exists := allServices[serviceName]; exists {
+						return verbMap, nil
+					}
+				}
+			}
+		}
+	}
+
+	verbResourceMap, err := fetchVerbResourceMap(serviceName, config)
+	if err != nil {
+		return nil, err
+	}
+
+	var allServices map[string]map[string][]string
+	if data, err := os.ReadFile(cacheFile); err == nil {
+		yaml.Unmarshal(data, &allServices)
+	}
+	if allServices == nil {
+		allServices = make(map[string]map[string][]string)
+	}
+
+	allServices[serviceName] = verbResourceMap
+
+	if err := os.MkdirAll(cacheDir, 0755); err == nil {
+		data, err := yaml.Marshal(allServices)
+		if err == nil {
+			os.WriteFile(cacheFile, data, 0644)
+		}
+	}
+
+	return verbResourceMap, nil
+}
+
+func handleLocalEnvironment(serviceName string) (map[string][]string, error) {
+	// TODO: check services
+	//if serviceName != "plugin" {
+	//	return nil, fmt.Errorf("only plugin service is supported in local environment")
+	//}
+
+	conn, err := grpc.Dial("localhost:50051", grpc.WithInsecure())
+	if err != nil {
+		return nil, fmt.Errorf("failed to connect to local plugin service: %v", err)
+	}
+	defer conn.Close()
+
+	ctx := context.Background()
+	refClient := grpcreflect.NewClient(ctx, grpc_reflection_v1alpha.NewServerReflectionClient(conn))
+	defer refClient.Reset()
+
+	services, err := refClient.ListServices()
+	if err != nil {
+		return nil, fmt.Errorf("failed to list local services: %v", err)
+	}
+
+	verbResourceMap := make(map[string][]string)
+	for _, s := range services {
+		// Skip grpc reflection services
+		if strings.HasPrefix(s, "grpc.") {
+			continue
+		}
+
+		// Handle plugin service
+		if serviceName == "plugin" && strings.Contains(s, ".plugin.") {
+			serviceDesc, err := refClient.ResolveService(s)
+			if err != nil {
+				continue
+			}
+
+			resourceName := s[strings.LastIndex(s, ".")+1:]
+			for _, method := range serviceDesc.GetMethods() {
+				verb := method.GetName()
+				if resources, ok := verbResourceMap[verb]; ok {
+					verbResourceMap[verb] = append(resources, resourceName)
+				} else {
+					verbResourceMap[verb] = []string{resourceName}
+				}
+			}
+			continue
+		}
+
+		// Handle other microservices
+		if strings.Contains(s, fmt.Sprintf("spaceone.api.%s.", serviceName)) {
+			serviceDesc, err := refClient.ResolveService(s)
+			if err != nil {
+				continue
+			}
+
+			resourceName := s[strings.LastIndex(s, ".")+1:]
+			for _, method := range serviceDesc.GetMethods() {
+				verb := method.GetName()
+				if resources, ok := verbResourceMap[verb]; ok {
+					verbResourceMap[verb] = append(resources, resourceName)
+				} else {
+					verbResourceMap[verb] = []string{resourceName}
+				}
+			}
+		}
+	}
+
+	return verbResourceMap, nil
+}
+
+func fetchVerbResourceMap(serviceName string, config *settings.Config) (map[string][]string, error) {
+	envConfig := config.Environments[config.Environment]
+	if envConfig.Endpoint == "" {
+		return nil, fmt.Errorf("endpoint not found in environment config")
+	}
+
+	var conn *grpc.ClientConn
+	var err error
+
+	if config.Environment == "local" {
+		endpoint := strings.TrimPrefix(envConfig.Endpoint, "grpc://")
+		conn, err = grpc.Dial(endpoint, grpc.WithInsecure())
+		if err != nil {
+			return nil, fmt.Errorf("connection failed: %v", err)
+		}
+	} else {
+		tlsConfig := &tls.Config{
+			InsecureSkipVerify: false,
+		}
+		apiEndpoint, _ := other.GetAPIEndpoint(envConfig.Endpoint)
+		identityEndpoint, hasIdentityService, err := other.GetIdentityEndpoint(apiEndpoint)
+
+		if !hasIdentityService {
+			// Get endpoints map first
+			endpointsMap, err := other.FetchEndpointsMap(apiEndpoint)
+			if err != nil {
+				return nil, fmt.Errorf("failed to fetch endpoints map: %v", err)
+			}
+
+			// Find the endpoint for the current service
+			endpoint, exists := endpointsMap[serviceName]
+			if !exists {
+				return nil, fmt.Errorf("endpoint not found for service: %s", serviceName)
+			}
+
+			// Parse the endpoint
+			parts := strings.Split(endpoint, "://")
+			if len(parts) != 2 {
+				return nil, fmt.Errorf("invalid endpoint format: %s", endpoint)
+			}
+
+			// Extract hostPort (remove the /v1 suffix if present)
+			hostPort := strings.Split(parts[1], "/")[0]
+
+			creds := credentials.NewTLS(tlsConfig)
+			conn, err = grpc.Dial(hostPort, grpc.WithTransportCredentials(creds))
+			if err != nil {
+				return nil, fmt.Errorf("connection failed: %v", err)
+			}
+		} else {
+			trimmedEndpoint := strings.TrimPrefix(identityEndpoint, "grpc+ssl://")
+			parts := strings.Split(trimmedEndpoint, ".")
+			if len(parts) < 4 {
+				return nil, fmt.Errorf("invalid endpoint format: %s", trimmedEndpoint)
+			}
+
+			// Replace 'identity' with the converted service name
+			parts[0] = format.ConvertServiceNameToEndpoint(serviceName)
+			serviceEndpoint := strings.Join(parts, ".")
+
+			creds := credentials.NewTLS(tlsConfig)
+			conn, err = grpc.Dial(serviceEndpoint, grpc.WithTransportCredentials(creds))
+			if err != nil {
+				return nil, fmt.Errorf("connection failed: %v", err)
+			}
+		}
+	}
+	defer conn.Close()
+
+	ctx := metadata.AppendToOutgoingContext(context.Background(), "token", envConfig.Token)
+	refClient := grpcreflect.NewClient(ctx, grpc_reflection_v1alpha.NewServerReflectionClient(conn))
+	defer refClient.Reset()
+
+	services, err := refClient.ListServices()
+	if err != nil {
+		return nil, fmt.Errorf("failed to list services: %v", err)
+	}
+
+	verbResourceMap := make(map[string][]string)
+	for _, s := range services {
+		if !strings.Contains(s, fmt.Sprintf(".%s.", serviceName)) {
+			continue
+		}
+
+		serviceDesc, err := refClient.ResolveService(s)
+		if err != nil {
+			continue
+		}
+
+		resourceName := s[strings.LastIndex(s, ".")+1:]
+		for _, method := range serviceDesc.GetMethods() {
+			verb := method.GetName()
+			if resources, ok := verbResourceMap[verb]; ok {
+				verbResourceMap[verb] = append(resources, resourceName)
+			} else {
+				verbResourceMap[verb] = []string{resourceName}
+			}
+		}
+	}
+
+	return verbResourceMap, nil
 }
 
 // watchResource monitors a resource for changes and prints updates
@@ -291,7 +532,7 @@ func printNewItems(items []map[string]interface{}) {
 		row := make([]string, len(headers))
 		for i, header := range headers {
 			if val, ok := item[header]; ok {
-				row[i] = formatTableValue(val)
+				row[i] = FormatTableValue(val)
 			}
 		}
 		tableData = append(tableData, row)
